LPS-92618 Sort

+++ modules/util/portal-tools-rest-builder/src/main/java/com/liferay/portal/tools/rest/builder/RESTBuilder.java
@@ -1009,8 +1009,8 @@ public class RESTBuilder {
 				Schema propertySchema = entry2.getValue();
 
 				if (Objects.equals(propertySchema.getType(), "number") &&
-					!Objects.equals(propertySchema.getFormat(), "float") &&
-					!Objects.equals(propertySchema.getFormat(), "double")) {
+					!Objects.equals(propertySchema.getFormat(), "double") &&
+					!Objects.equals(propertySchema.getFormat(), "float")) {
 
 					StringBuilder sb = new StringBuilder();
